requirements.txt
aiogram==3.7.0
SQLAlchemy[asyncio]==2.0.31
asyncpg==0.29.0
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow>=10.4.0